+++ sentry-android-okhttp/src/test/java/io/sentry/android/okhttp/SentryOkHttpInterceptorTest.kt
@@ -253,8 +253,8 @@ class SentryOkHttpInterceptorTest {
         verify(fixture.hub).addBreadcrumb(
             check<Breadcrumb> {
                 assertEquals("http", it.type)
-                assertEquals(13L, it.data["response_body_size"])
-                assertEquals(12L, it.data["request_body_size"])
+                assertEquals(13L, it.data["http.response_content_length"])
+                assertEquals(12L, it.data["http.request_content_length"])
             },
             anyOrNull()
         )